Made pretties and no more warnings
git-svn-id: file:///home/nbingham/backup/private-repos/trunk/Software/Work/chp@99 404bc64d-9d1b-498e-bf1b-9a6e5dac48b1

+++ src/state.cpp
@@ -103,7 +103,7 @@ state &state::operator>>=(int n)
 	return *this;
 }
 
-state state::operator[](int i)
+state state::operator[](size_t i)
 {
 	stringstream str;
 	string s;
